fix(cli): repair gateway lifecycle checks
