+++ src/cli/daemon-cli/lifecycle.ts
@@ -29,16 +29,14 @@ import {
   formatExternalSupervisorActionRequired,
   isGatewayExternallySupervised,
 } from "../../infra/gateway-supervision.js";
-import type { SafeGatewayRestartRequestResult } from "../../infra/restart-coordinator.js";
 import {
   clearGatewayRestartIntentSync,
   type GatewayRestartIntent,
   writeGatewayRestartIntentSync,
 } from "../../infra/restart-intent.js";
 import { resolveGatewayRestartDeferralTimeoutMs } from "../../infra/restart.js";
-import { defaultRuntime, writeRuntimeJson } from "../../runtime.js";
+import { defaultRuntime } from "../../runtime.js";
 import { formatCliCommand } from "../command-format.js";
-import { parseDurationMs } from "../parse-duration.js";
 import {
   isTerminalInteractive,
   NON_INTERACTIVE_GATEWAY_STOP_MESSAGE,
@@ -54,6 +52,10 @@ import {
   runServiceStop,
   runServiceUninstall,
 } from "./lifecycle-core.js";
+import {
+  requestSafeGatewayRestart,
+  resolveGatewayRestartIntentOptions,
+} from "./lifecycle-safe-restart.js";
 import { createDaemonActionContext, createNullWriter } from "./response.js";
 import {
   DEFAULT_RESTART_HEALTH_ATTEMPTS,
@@ -222,21 +224,6 @@ async function stopGatewayWithoutServiceManager(port: number) {
   };
 }
 
-function resolveGatewayRestartIntentOptions(
-  opts: DaemonLifecycleOptions,
-): GatewayRestartIntent | undefined {
-  if (opts.force && opts.wait !== undefined) {
-    throw new Error("--force cannot be combined with --wait");
-  }
-  if (opts.force) {
-    return { force: true };
-  }
-  if (opts.wait !== undefined) {
-    return { waitMs: parseDurationMs(opts.wait) };
-  }
-  return undefined;
-}
-
 async function resolveRestartListenerHealthWait(
   restartIntent: GatewayRestartIntent | undefined,
 ): Promise<{
@@ -271,64 +258,6 @@ async function resolveRestartListenerHealthWait(
     waitIndefinitelyForPreviousOwner: false,
     timeoutSeconds: Math.round((attempts * POST_RESTART_HEALTH_DELAY_MS) / 1000),
   };
-}
-
-function formatSafeRestartWarnings(result: SafeGatewayRestartRequestResult): string[] | undefined {
-  if (result.preflight.blockers.length === 0) {
-    return undefined;
-  }
-  return [result.preflight.summary];
-}
-
-async function requestSafeGatewayRestart(opts: DaemonLifecycleOptions): Promise<boolean> {
-  if (opts.force) {
-    throw new Error("--safe cannot be combined with --force; omit --safe to force restart now");
-  }
-  if (opts.wait !== undefined) {
-    throw new Error("--safe cannot be combined with --wait; safe restart uses gateway deferral");
-  }
-  const skipDeferral = opts.skipDeferral === true;
-  const params: { reason: string; skipDeferral?: true } = { reason: "gateway.restart.safe" };
-  if (skipDeferral) {
-    params.skipDeferral = true;
-  }
-  const result = await callGatewayCli<SafeGatewayRestartRequestResult>({
-    method: "gateway.restart.request",
-    params,
-    timeoutMs: 10_000,
-  });
-  appendGatewayLifecycleAudit({
-    action: "restart",
-    source: "safe-rpc",
-    mode: result.status,
-    pid: result.restart.pid,
-  });
-  const message =
-    result.status === "coalesced"
-      ? "safe restart request joined an existing pending gateway restart"
-      : result.status === "deferred"
-        ? "safe restart requested; gateway will restart after active work drains " +
-          "(bounded by gateway.reload.deferralTimeoutMs; may force after timeout expires)"
-        : skipDeferral
-          ? "safe restart requested; gateway bypassing active-work deferral"
-          : "safe restart requested; gateway will restart momentarily";
-  const payload = {
-    ok: true,
-    result: result.status,
-    message,
-    preflight: result.preflight,
-    restart: result.restart,
-    warnings: formatSafeRestartWarnings(result),
-  };
-  if (opts.json) {
-    writeRuntimeJson(defaultRuntime, payload);
-  } else {
-    defaultRuntime.log(message);
-    if (result.preflight.blockers.length > 0) {
-      defaultRuntime.log(theme.warn(result.preflight.summary));
-    }
-  }
-  return true;
 }
 
 async function signalGatewayRestart(
